+++ src/section-heading/index.ts
@@ -6,6 +6,7 @@ import { registerBlockType } from '@wordpress/blocks';
 
 import Edit from './edit';
 import Save from './save';
+import deprecated from './deprecated';
 import metadata from './block.json';
 
 import './style.scss';
@@ -17,4 +18,5 @@ registerBlockType(metadata.name as any, {
 	...metadata,
 	edit: Edit,
 	save: Save,
+	deprecated,
 } as any);